Update director.py
it is working but have errors: I am not able to get the ramdon.randint in the CARD Class that is why I used it inside the Director Class, we have 2 variables for random randint and there is possibility that the current number and next number are the same, the elif self.ask.user == "n" does not working(the game continues). I like to continue working but I have to rest for now, I hope someone will continue to fix the errors. We only have almost 2 days before due date. Thanks. Happy Coding!